# api/openapi.yaml
openapi: 3.1.0
info:
  title: Puerta Abierta API
  version: 1.0.0
  summary: STR operations API for Paraguay-focused owners and operators
  description: |
    REST API contract for multi-tenant short-term rental operations:
    organizations, properties, reservations, tasks, finance, and reporting.
servers:
  - url: https://api.puertaabierta.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /reservations:
    get:
      tags: [Reservations]
      summary: List reservations
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: unit_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReservationStatus"
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Reservation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Reservation"
                required: [data]
    post:
      tags: [Reservations]
      summary: Create reservation
      description: |
        Creates a reservation. API must enforce no overlapping active reservations
        for the same unit (`pending`, `confirmed`, `checked_in`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "409":
          description: Overlap or duplicate external reservation id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reservations/{reservation_id}:
    get:
      tags: [Reservations]
      summary: Get reservation by id
      parameters:
        - $ref: "#/components/parameters/ReservationIdPath"
      responses:
        "200":
          description: Reservation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Reservations]
      summary: Update reservation
      parameters:
        - $ref: "#/components/parameters/ReservationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationInput"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{reservation_id}/status:
    post:
      tags: [Reservations]
      summary: Transition reservation status
      parameters:
        - $ref: "#/components/parameters/ReservationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/ReservationStatus"
                reason:
                  type: string
              required: [status]
      responses:
        "200":
          description: Status transitioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "422":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: assigned_user_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                required: [data]
    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{task_id}:
    get:
      tags: [Tasks]
      summary: Get task by id
      parameters:
        - $ref: "#/components/parameters/TaskIdPath"
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Tasks]
      summary: Update task
      parameters:
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{task_id}/complete:
    post:
      tags: [Tasks]
      summary: Complete task
      parameters:
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                completion_notes:
                  type: string
      responses:
        "200":
          description: Task marked done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrgIdQuery:
      name: org_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
    ReservationIdPath:
      name: reservation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskIdPath:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    UUID:
      type: string
      format: uuid
    Currency:
      type: string
      enum: [PYG, USD]
...
    ReservationStatus:
      type: string
      enum: [pending, confirmed, checked_in, checked_out, cancelled, no_show]
    TaskType:
      type: string
      enum: [cleaning, maintenance, check_in, check_out, inspection, custom]
    TaskStatus:
      type: string
      enum: [todo, in_progress, done, cancelled]
    PriorityLevel:
      type: string
      enum: [low, medium, high, urgent]
...
    PaymentMethod:
      type: string
      enum: [bank_transfer, cash, card, qr, other]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]
...
    Reservation:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        unit_id:
          $ref: "#/components/schemas/UUID"
        listing_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        channel_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        guest_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        external_reservation_id:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ReservationStatus"
        source:
          type: string
        check_in_date:
          type: string
          format: date
        check_out_date:
          type: string
          format: date
        adults:
          type: integer
        children:
          type: integer
        infants:
          type: integer
        pets:
          type: integer
        currency:
          $ref: "#/components/schemas/Currency"
        total_amount:
          type: number
        amount_paid:
          type: number
        owner_payout_estimate:
          type: number
      required:
        [id, organization_id, unit_id, status, source, check_in_date, check_out_date, currency, total_amount, amount_paid, owner_payout_estimate]
...
    Task:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        reservation_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        assigned_user_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        type:
          $ref: "#/components/schemas/TaskType"
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        title:
          type: string
        description:
          type: string
          nullable: true
        due_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
      required: [id, organization_id, type, status, priority, title]
...
    CreateReservationInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        unit_id:
          $ref: "#/components/schemas/UUID"
        listing_id:
          $ref: "#/components/schemas/UUID"
        channel_id:
          $ref: "#/components/schemas/UUID"
        guest_id:
          $ref: "#/components/schemas/UUID"
        external_reservation_id:
          type: string
        source:
          type: string
          default: manual
        check_in_date:
          type: string
          format: date
        check_out_date:
          type: string
          format: date
        adults:
          type: integer
        children:
          type: integer
        infants:
          type: integer
        pets:
          type: integer
        currency:
          $ref: "#/components/schemas/Currency"
        nightly_rate:
          type: number
        cleaning_fee:
          type: number
        tax_amount:
          type: number
        extra_fees:
          type: number
        discount_amount:
          type: number
        total_amount:
          type: number
        payment_method:
          $ref: "#/components/schemas/PaymentMethod"
        notes:
          type: string
      required:
        [organization_id, unit_id, check_in_date, check_out_date, total_amount]
    UpdateReservationInput:
      type: object
      properties:
        guest_id:
          $ref: "#/components/schemas/UUID"
        status:
          $ref: "#/components/schemas/ReservationStatus"
        payment_method:
          $ref: "#/components/schemas/PaymentMethod"
        amount_paid:
          type: number
        notes:
          type: string
...
    CreateTaskInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        property_id:
          $ref: "#/components/schemas/UUID"
        unit_id:
          $ref: "#/components/schemas/UUID"
        reservation_id:
          $ref: "#/components/schemas/UUID"
        assigned_user_id:
          $ref: "#/components/schemas/UUID"
        type:
          $ref: "#/components/schemas/TaskType"
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        title:
          type: string
        description:
          type: string
        due_at:
          type: string
          format: date-time
      required: [organization_id, title]
    UpdateTaskInput:
      type: object
      properties:
        assigned_user_id:
          $ref: "#/components/schemas/UUID"
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        title:
          type: string
        description:
          type: string
        due_at:
          type: string
          format: date-time
...